+++ marketing-service/src/main/java/net/devk/marketing/service/requirements/RequirementService.java
@@ -4,9 +4,9 @@ import net.devk.marketing.service.requirements.dto.CreateNewRequirementResponseD
 
 public interface RequirementService {
 
-	public CreateNewRequirementResponseDTO createRequirement(Long customerId, Long targetMemberId, Integer estimatedValue,
+	public CreateNewRequirementResponseDTO createRequirement(Long customerId, Long targetMemberId, Long estimatedValue,
 			String description);
 	
-	public void assignRequirement(Long requirementId, Long personnelId , Long assignedStatusTypeId,Integer realValue);
+	public void assignRequirement(Long requirementId, Long personnelId , Long assignedStatusTypeId,Long realValue);
 
 }
