+++ marketing-service/src/main/java/net/devk/marketing/service/requirements/dto/CreateNewRequirementResponseDTO.java
@@ -10,8 +10,8 @@ import lombok.RequiredArgsConstructor;
 public class CreateNewRequirementResponseDTO {
 
 	private final Long id;
-	private final Integer estimatedValue;
-	private final Integer realValue;
+	private final Long estimatedValue;
+	private final Long realValue;
 	private final String description;
 	private final Date estimatedValueEditDate;
 	private final Date estimatedValueRegisterDate;
